ai-service/requirements.txt
groq==0.4.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
PyMuPDF==1.23.6
python-pptx==0.6.22
httpx==0.25.1
python-dotenv==1.0.1
